Removing "go mod vendor" from test script
Test script was running `go mod vendor` with golang 1.13 from the default Travis container. This removes the "explicit" tags in modules.txt [1], causing the build to fail with "	collectd.org@v0.3.0: is explicitly required in go.mod, but not marked as explicit in vendor/modules.txt" when it is later built for real with golang 1.14 [2]. This script should probably run tests using the same version of golang we will eventually build with.  I'm removing this line because we do the `go mod vendor` before committing now. 

[1] https://github.com/infrawatch/sg-core/blob/master/vendor/modules.txt#L2
[2] https://travis-ci.org/github/infrawatch/sg-core/builds/743489014#L1237