Silently drop the return codes during unit testing
TODO: we will fix this during the code refactoring phase over the next sprint or
so. For now let's get linting to pass so that it can check for any new issues we
introduce during the refactoring. TODO items have been added.

We need to review these unit tests anyways since they are currently failing.