Build test plugins for upstream distribution
Build the test plugins dummy-* for upstream builds so that when testing
with smart-gateway-operator by default the plugins exist to simulate
workloads without needing a source connected.

Changes will be used as part of the smart-gateway-operator example
CustomResource and allow for additional molecule testing in the future
for that Operator.