Build test plugins for upstream distribution (#95)
Build the test plugins dummy-* for upstream builds so that when testing
with smart-gateway-operator by default the plugins exist to simulate
workloads without needing a source connected.

Changes will be used as part of the smart-gateway-operator example
CustomResource and allow for additional molecule testing in the future
for that Operator.

Cherry picked from commit 5b9695b392d33baa5702d83338fff6163ea8b45f