Update GITHUB_REF for GITHUB_HEAD_REF
GITHUB_HEAD_REF points to the head ref or source branch
of the pull request in a workflow run.

GITHUB_REF points to the pull request merge branch.
The ref given is fully-formed, meaning that for pull
requests it is refs/pull/<pr_number>/merge

We should use the former to achieve the retrieving
a branch with the same name