Install and run node_exporter
In order to collect metrics from the compute node in DevStack
environment we need to install node_exporter on each node and add these
nodes to scrape_configs in prometheus.yaml.

This pr automates installation and running of node exporter on the node.

In order to add it to scrape_configs, we need to add the compute ip and
port to PROMETHEUS_CUSTOM_SCRAPE_TARGETS.

Signed-off-by: Chandan Kumar (raukadah) <raukadah@gmail.com>